Add flag_item
Added the flag_item funtion is scripts.js that uses ajax to run the _flag_item route. Also updated buy/item.html to check if a user is logged in or not for both flagging and saving listings. User is alerted to login for saves. Flagging can be done anonymously. Currently flagging does not track who flags what but this capacity can be implemented easily in the views.

Time spent on issue: 1:30