Add admin blueprint
Admin blueprint has basic admin portal. It's route is login_required. If the user is logged in but not an admin (the default) the route renders the market.buy template. If the user __is__ an admin, they see the admin portal.

The portal--as it is now--contains only a list of flagged items. There is no admin capacity to delete listings currently.

Time Spent 1:00