Add admin delete route
Added a new inspect route that renders an inspect template. Then admin can delete post. Would like to turn the inspect template into a form so the admin can edit Title, Body, and Pictures as well as send the user a warning, ban the user, etc. on a post-by post basis.

Time Spent 1:00